Move `getAuthConfig` out of `start-proxy.ts`
